Refactor rewinding logic into new Rewinding module
Moved DeltaEncoder, FixedSizeRingbuffer, and RewindController from YAGEFrontend to a new src/Rewinding module and created a dedicated Rewinding static library. Updated project files, solution files, and dependencies to use the new module. Added new tests for rewinding functionality and improved test file handling. SerializationView size is now uint64_t for better compatibility. Minor code cleanups and interface adjustments to support the new structure.

+++ src/YAGECore/Source/Serialization.cpp
@@ -235,7 +235,7 @@ void SerializationFactory::WriteChunkHeader(uint32_t writeDataSize, const ChunkI
 	m_serializedChunks++;
 }
 
-DeserializationFactory::DeserializationFactory(SerializationParameters parameters, const uint8_t* buffer, const uint32_t size)
+DeserializationFactory::DeserializationFactory(SerializationParameters parameters, const uint8_t* buffer, const uint64_t size)
 	: m_parameters(parameters)
 	, m_finished(false)
 {
